# openapi.yaml
openapi: 3.0.3
info:
  title: Media Analyzer API
  description: |
    Comprehensive API for analyzing social media content, particularly Instagram Reels, with advanced capabilities including:
    - **Content Analysis**: Full compliance checking and brand safety analysis
    - **Keyword Extraction**: Advanced semantic keyword extraction with sentiment analysis
    - **Media Processing**: Audio transcription (ASR) and optical character recognition (OCR)
    - **Named Entity Recognition**: ML-based entity extraction and semantic similarity
    - **Multilingual Support**: English, Hindi, and Hinglish content processing
    
    ## System Architecture
    
    The system consists of three main components:
    
    1. **Main API Server (Node.js/Fastify)** - Port 3000
    2. **Python Worker (FastAPI)** - Port 8000
    3. **Database & Queue System** - SQLite with Redis
    
    ## Authentication
    
    All protected endpoints require JWT authentication via Bearer token in the Authorization header.
    
  version: 1.0.0
  contact:
    name: Media Analyzer API Support
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000
    description: Main API Server (Development)
  - url: http://localhost:8000
    description: Python Worker Server (Development)
  - url: https://api.media-analyzer.com
    description: Production Server
...
paths:
...
  /asr:
    post:
      tags:
        - Worker - Media Processing
      summary: Speech-to-text transcription
      description: |
        Automatic Speech Recognition (ASR) using Whisper model.
        
        **Features:**
        - Multiple language support
        - Audio preprocessing (noise reduction, normalization)
        - Text post-processing (capitalization, punctuation)
        - Configurable model size and compute type
      operationId: asr
      servers:
        - url: http://localhost:8000
          description: Python Worker Server
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
              properties:
                file:
                  type: string
                  format: binary
                  description: Audio file to transcribe
                language:
                  type: string
                  description: Language hint (e.g., "en", "hi", "es")
                  example: en
                enable_preprocessing:
                  type: boolean
                  description: Enable audio preprocessing
                preprocessing_level:
                  type: string
                  enum: [minimal, standard, aggressive]
                  description: Preprocessing level
                enable_postprocessing:
                  type: boolean
                  description: Enable text post-processing
      responses:
        '200':
          description: Transcription completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ASRResponse'
        '500':
          description: ASR processing failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /ocr:
    post:
      tags:
        - Worker - Media Processing
      summary: Optical character recognition
      description: Extracts text from image frames using Tesseract OCR
      operationId: ocr
      servers:
        - url: http://localhost:8000
          description: Python Worker Server
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - files
              properties:
                files:
                  type: array
                  items:
                    type: string
                    format: binary
                  description: Image files for OCR processing
      responses:
        '200':
          description: OCR completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OCRResponse'
        '500':
          description: OCR processing failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /download-instagram:
    post:
      tags:
        - Worker - Media Processing
      summary: Download Instagram Reel
      description: |
        Downloads Instagram Reel video with metadata.
        
        **Cookie Options:**
        - Browser cookies (chrome, firefox, safari, edge, opera, brave)
        - Cookies file (Netscape format)
      operationId: downloadInstagram
      servers:
        - url: http://localhost:8000
          description: Python Worker Server
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InstagramDownloadRequest'
            examples:
              basic:
                summary: Basic download
                value:
                  url: https://www.instagram.com/reel/ABC123DEF456/
              with_browser_cookies:
                summary: With browser cookies
                value:
                  url: https://www.instagram.com/reel/ABC123DEF456/
                  browser_cookies: chrome
              with_cookies_file:
                summary: With cookies file
                value:
                  url: https://www.instagram.com/reel/ABC123DEF456/
                  cookies_file: /path/to/cookies.txt
      responses:
        '200':
          description: Download completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InstagramDownloadResponse'
  /ner:
    post:
      tags:
        - Worker - NLP
      summary: Named Entity Recognition
      description: |
        Extracts named entities from text using ML-based NER models.
        
        **Supported Languages:** English, Hindi, Hinglish
        
        **Entity Types:** Person, Organization, Location, Date, Time, Money, etc.
      operationId: ner
      servers:
        - url: http://localhost:8000
          description: Python Worker Server
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NERRequest'
            examples:
              english:
                summary: English text
                value:
                  text: "Apple Inc. is headquartered in Cupertino, California."
                  language: en
                  include_relationships: true
              hindi:
                summary: Hindi text
                value:
                  text: "मुंबई भारत का सबसे बड़ा शहर है।"
                  language: hi
                  include_relationships: true
      responses:
        '200':
          description: NER completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NERResponse'
        '500':
          description: NER processing failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /semantic-similarity:
    post:
      tags:
        - Worker - NLP
      summary: Semantic similarity analysis
      description: |
        Computes semantic similarity between keywords and clusters them.
        
        **Features:**
        - Multilingual embeddings
        - Keyword clustering
        - Similarity matrix computation
        - Grouped keyword recommendations
      operationId: semanticSimilarity
      servers:
        - url: http://localhost:8000
          description: Python Worker Server
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SemanticSimilarityRequest'
            examples:
              basic:
                summary: Basic similarity
                value:
                  keywords: ["fashion", "style", "clothing", "trendy", "design"]
                  language: en
                  cluster: true
      responses:
        '200':
          description: Similarity computed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SemanticSimilarityResponse'
        '500':
          description: Processing failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

# ==================== Components ====================
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /auth/login or /auth/demo-token
  schemas:
...
    ASRResponse:
      type: object
      properties:
        language:
          type: string
          example: en
        segments:
          type: array
          items:
            type: object
            properties:
              tStart:
                type: number
                format: float
              tEnd:
                type: number
                format: float
              text:
                type: string
        timing:
          type: number
          description: Processing time in milliseconds
        preprocessing_enabled:
          type: boolean
        preprocessing_level:
          type: string
          enum: [minimal, standard, aggressive]
        postprocessing_enabled:
          type: boolean
        model_size:
          type: string
        compute_type:
          type: string
    OCRResponse:
      type: object
      properties:
        frames:
          type: array
          items:
            type: object
            properties:
              t:
                type: integer
              boxes:
                type: array
                items:
                  type: object
                  properties:
                    box:
                      type: array
                      items:
                        type: integer
                      minItems: 4
                      maxItems: 4
                    text:
                      type: string
        timing:
          type: number
          description: Processing time in milliseconds
    InstagramDownloadRequest:
      type: object
      required:
        - url
      properties:
        url:
          type: string
          format: uri
          example: https://www.instagram.com/reel/ABC123DEF456/
        output_path:
          type: string
          nullable: true
        browser_cookies:
          type: string
          enum: [chrome, firefox, safari, edge, opera, brave]
          nullable: true
        cookies_file:
          type: string
          nullable: true
    InstagramDownloadResponse:
      type: object
      properties:
        success:
          type: boolean
        video_path:
          type: string
          nullable: true
        caption:
          type: string
          nullable: true
        username:
          type: string
          nullable: true
        duration:
          type: number
          nullable: true
        error:
          type: string
          nullable: true
    NERRequest:
      type: object
      required:
        - text
      properties:
        text:
          type: string
          example: "Apple Inc. is headquartered in Cupertino, California."
        language:
          type: string
          default: en
          example: en
        include_relationships:
          type: boolean
          default: true
    NERResponse:
      type: object
      properties:
        entities:
          type: object
          additionalProperties:
            type: array
            items:
              type: object
        relationships:
          type: array
          items:
            type: object
        metadata:
          type: object
        timing:
          type: number
          description: Processing time in milliseconds
    SemanticSimilarityRequest:
      type: object
      required:
        - keywords
      properties:
        keywords:
          type: array
          items:
            type: string
          example: ["fashion", "style", "clothing", "trendy", "design"]
        language:
          type: string
          default: en
          example: en
        cluster:
          type: boolean
          default: true
    SemanticSimilarityResponse:
      type: object
      properties:
        clusters:
          type: array
          items:
            type: object
        similarity_matrix:
          type: array
          items:
            type: array
            items:
              type: number
        grouped_keywords:
          type: object
          additionalProperties:
            type: array
            items:
              type: object
        embeddings_shape:
          type: array
          items:
            type: integer
        metadata:
          type: object
        timing:
          type: number
        error:
          type: string
          nullable: true

    # ==================== Common Schemas ====================
...
    ErrorResponse:
      type: object
      properties:
        code:
          type: string
          example: INTERNAL_ERROR
        message:
          type: string
          example: An error occurred
        details:
          type: string

  # ==================== Reusable Responses ====================
...